requirements.txt
fastapi
uvicorn
pydantic
kafka-python-ng